requirements.txt
fastapi
uvicorn
supabase
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
PyJWT
httpx
razorpay